disable webpack bit angular plugin
This plugin was used to run ngcc on both local and capsule folders but it could cause conflict issues with the ones integrated into angular cli. Since it was just a precaution to make sure that everything was precompiled (but shouldn't be necessary anyway), we are disabling it to avoid this kind of issues.